+++ docs/Whats-New.md
@@ -171,6 +171,7 @@ New:
 - Implement `DecloakToFire` (by ZivDero)
 - Implement `CellSpread` and `PercentAtMax` (by ZivDero)
 - Implement `ScorchChance`, `CraterChance` and `CellAnimChance` (by ZivDero)
+- Implement `Explosive` for animations (by ZivDero, Rampastring)
 
 
 Vanilla fixes: